# dev/backend/docs/openapi/condohub.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CondoHub Backend API",
    "version": "1.0.0",
    "description": "API do CondoHub Backend (Dev 1 + Dev 2, refletindo endpoints implementados)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local"
    },
    {
      "url": "https://api-dev.condohub.local",
      "description": "Exemplo ambiente dev"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/{instanceKey}/auth/register-with-invite": {
      "post": {
        "tags": [
          "Auth Tenant"
        ],
        "summary": "Register with invite",
        "description": "Rota pública tenant para cadastro via fluxo novo de invite por código/QR. Retorna AuthResponse canônico.",
        "operationId": "registerWithInvite",
        "security": [],
        "parameters": [
          {
            "$ref": "#/components/parameters/instanceKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterWithInviteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authenticated after invite signup",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation or invite/signup context error (codes: VALIDATION_ERROR, INVALID_SIGNUP_TOKEN, SIGNUP_TOKEN_EXPIRED, INVALID_INVITE_CONTEXT, INVITE_ALREADY_USED, INVITE_REVOKED, INVITE_EXPIRED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "$ref": "#/components/responses/InstanceSuspendedError"
          },
          "404": {
            "$ref": "#/components/responses/TenantNotFoundError"
          },
          "409": {
            "$ref": "#/components/responses/UserEmailConflictError"
          },
          "429": {
            "$ref": "#/components/responses/RateLimitExceededError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "instanceKey": {
        "name": "instanceKey",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        },
        "description": "Chave da instância (tenant)"
      },
...
    },
    "responses": {
...
      "TenantNotFoundError": {
        "description": "Tenant not found (code: TENANT_NOT_FOUND)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": true,
              "code": "TENANT_NOT_FOUND",
              "message": "Instance not found",
              "details": {}
            }
          }
        }
      },
      "InstanceSuspendedError": {
        "description": "Instance suspended (code: INST_SUSPENDED)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": true,
              "code": "INST_SUSPENDED",
              "message": "This instance is suspended",
              "details": {}
            }
          }
        }
      },
...
      "RateLimitExceededError": {
        "description": "Rate limit exceeded (code: RATE_LIMIT_EXCEEDED)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": true,
              "code": "RATE_LIMIT_EXCEEDED",
              "message": "Too many requests, please try again later",
              "details": {}
            }
          }
        }
      },
      "UserEmailConflictError": {
        "description": "User email conflict (code: USER_EMAIL_CONFLICT)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": true,
              "code": "USER_EMAIL_CONFLICT",
              "message": "Email already exists in this instance",
              "details": {}
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "code",
          "message",
          "details"
        ],
        "properties": {
          "error": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "additionalProperties": true
          }
        },
        "example": {
          "error": true,
          "code": "SOME_CODE",
          "message": "Human message",
          "details": {}
        }
      },
      "Role": {
        "type": "string",
        "enum": [
          "ADMIN_GLOBAL",
          "SINDICO_ADMIN",
          "FUNC_ENTREGAS",
          "FUNC_MANUTENCAO",
          "MORADOR"
        ]
      },
      "AuthUser": {
        "type": "object",
        "required": [
          "id",
          "instanceId",
          "unitId",
          "roles",
          "name",
          "email",
          "phone"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "instanceId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "instanceKey": {
            "type": "string",
            "nullable": true,
            "description": "Chave pública da instância tenant. Retornada em logins tenant/global quando aplicável; ausente ou null para admin global."
          },
          "unitId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "AuthResponse": {
        "type": "object",
        "required": [
          "accessToken",
          "refreshToken",
          "expiresInSec",
          "user"
        ],
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresInSec": {
            "type": "integer",
            "enum": [
              300
            ]
          },
          "user": {
            "$ref": "#/components/schemas/AuthUser"
          }
        }
      },
...
      "RegisterWithInviteRequest": {
        "type": "object",
        "required": [
          "signupToken",
          "name",
          "email",
          "password"
        ],
        "properties": {
          "signupToken": {
            "type": "string",
            "minLength": 1
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 256
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 40
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}